fixed config file not found error

+++ scripts/main.py
@@ -6,6 +6,7 @@ import sys
 import traceback
 import yaml
 import os
+from mpi4py import MPI
 
 from btx.misc.shortcuts import AttrDict
 from scripts.tasks import *
@@ -20,21 +21,25 @@ def main():
         config = AttrDict(yaml.safe_load(config_file))
         #TODO: check required arguments in config dictionary here.
 
-    # Create output directory.
-    try:
-        os.makedirs(config.setup.root_dir, exist_ok=True)
-    except:
-        print(f"Error: cannot create root path.") 
-        return -1 
+    comm = MPI.COMM_WORLD
+    rank = comm.Get_rank()
+    if rank == 0:
+        # Create output directory.
+        try:
+            os.makedirs(config.setup.root_dir, exist_ok=True)
+        except:
+            print(f"Error: cannot create root path.") 
+            return -1 
 
-    path_plus_file = config.setup.root_dir + '/' + config_filepath.split('/')[-1]
-    if os.path.exists(path_plus_file):
-        os.remove(path_plus_file)
+        path_plus_file = config.setup.root_dir + '/' + config_filepath.split('/')[-1]
+        if os.path.exists(path_plus_file):
+            os.remove(path_plus_file)
         
-    # Copy config file to output directory.
-    shutil.copy2(config_filepath, config.setup.root_dir)
-    # Call 'task' function if it exists.
-        
+        # Copy config file to output directory.
+        shutil.copy2(config_filepath, config.setup.root_dir)
+        # Call 'task' function if it exists.
+    comm.Barrier()
+
     try:
         globals()[task]
     except Exception as e:
